rename and move Sideboardline
